Fix day filter bug
A bug still existed when the user attempts to filter leaves by year,
month AND day. The year and month are not correctly accounted for when a
day is specified. This is now fixed by calculating the endDate for
filtering from the startDate using the Moment library, as opposed to
implementing complex logic myself.